Use custom pkg cert
It turns out, we cannot simply replace the default cert because then pip
installs will fail because they expect the ca_root_nss cert.

however, it transpires that newer versions of pkg cannot bootstrap
themselves from a custom repo without setting the appropriate
environment variables on its first call (subsequent calls then honour
the pkg.conf settings)

That's why we need a custom version of the startup script that ships
with bsdploy (which does not assume a custom repo) and inject the ENV
settings)

Note that we do not template this startup script (i.o. the value is
hardcoded and will need to be adjusted for scenarios that change the
``jails_pkg_url`` setting.