Allow specifying a theme package
If set, all templates and the static route will be looked up using that
package.

This allows for a python package to provide a complete custom look
without using the diazo mechanism by overriding the jinja templates.

note that the theme package must be a python installable egg and
contain a ``templates`` folder that provides definitions for *all* the
templates that the briefkasten defines.