FIX: also pin pytest-flakes
the reason is, that newer versions of it also require a newwer version
of pytest itself, which we in turn pin, because newer version of that
produce *a lot* of errors in regards to the usage of packageresources
which is used for loading jinja and chameleon templates.

I tried using the very latest version of pytest but it still exhibits
that behaviour, so for now we simply also pin pytest-flakes.