feat(mail): allow to store the smtp password encrypted
This uses the `ploy_crypted` feature so the password can be
stored outside the playbook, i.e. in a "config" repository.

Note that in case `ploy` crashes with `ERROR! A worker was
found in a dead state` setting the environment variable
`OBJC_DISABLE_INITIALIZE_FORK_SAFETY` to `YES` might help.

See https://github.com/ployground/ploy_ansible#ansible-lookup-plugins
and https://stackoverflow.com/a/52230415